Beta Testing - Update 1.4
This update is called the "Open for Beta Testing" update. The site has been furnished for allowing beta testers to test out the website before launch.

Changelog:
- furnishing
- added a thank you message that took me 2 hours to code in
- google analytics tag added to derive data from beta tests.
- landscaping update (landscape bugs fixed
- blog page alert
- some minor bug fixes and code cleanup

Note: this is a beta testing phase, there are bugs that remain to be squashed, and there is filler text to be changed. Beta testing will be live for maximum of 30 days before shutdown.

Thank you for reading!
- Priyanshu (@probablyliquid)